Numerous changes to include PiGraphs in PreesmScenario:
-Changed TimingManager.getTimingOrDefault method to return a Timing object and not a long, since a Timing can no be chracterized by an expression
-Added a TimingManager.setTiming method which accept a String as value rather as a long
-Added a PiGraph attribute to ScenarioParser to handle PISDF algorithms, and modified ScenarioParser to handle this PiGraph
-Modified TimingsPage to add columns to the timings table (parsing, evaluation, variables) and change the time column to an expression one
-Modified SDFTableLabelProvider to provide text and images for Timings, based on PiScenario

+++ plugins/org.ietr.preesm.core/src/org/ietr/preesm/core/scenario/Timing.java
@@ -242,7 +242,7 @@ public class Timing {
 			for (String parameter : parametersValues.keySet())
 				jep.addVariable(parameter, parametersValues.get(parameter));
 			jep.parse(stringValue);
-			long result = (long) jep.evaluate();
+			long result = ((Double) jep.evaluate()).longValue();
 			this.time = result;
 			this.isEvaluated = true;
 		} catch (JepException e) {
